build: fix fluentui imports for webpack
The latest version of Fluent UI specifies a list of exported modules in its package.json. Only these exports can be safely imported. With SPFx 1.19 and Webpack 5, using other exports will crash while creating the bundle.

This PR updates the imports throughout the project to use the correct exports, which will allow the project to be used with the latest SPFx toolchain.